rename: 폴더구조 변경 (#182)
